Introduce AgentClusterID to serialized messages
There are restrictions to sharing SharedArrayBuffer (and WebAssembly
Modules) between execution contexts; for example, you cannot share a
SharedArrayBuffer between a Window and the Service Worker it created.
The rules are described here:
https://html.spec.whatwg.org/multipage/webappapis.html#integration-with-the-javascript-agent-cluster-formalism

This CL begins to model some of this behavior by introducing an
`AgentClusterID` (an alias of `base::UnguessableToken`) that can be
queried from an `ExecutionContext`.

Bug: chromium:798572
Change-Id: Ie70cce4cbd0ebd04d8d270d66f59690caf7f616a