Introduced blink::mojom::WebUsbService with restricted methods.
This CL introduced blink::mojom::WebUsbService implemented
by WebUsbServiceImpl in chrome/browser/usb which used to be
WebDeviceManager.
Paralleling with device::mojom::UsbDeviceManager, the new interface
aims to simplify the interactions with Blink, and a meaningless
argument of GetDevices() method has been reduced in this CL.
device::mojom::UsbChooserService will be merged to this interface in
subsequent CLs.

BUG=699790

Change-Id: I891c56460425c20d2e8d067f9335091f6945ae7d